Add route monitoring functionality
Added functionality to wait for routes. Was supposed to help when using
web_updater in case DNS server was down. Unfortunately, if interface is
down then NetworkManager blanks /etc/resolv.conf so can't find out what
the DNS server addresses are at that time. Could find them in advance,
or retry a set number of times on any route change.